% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_gerda_web.R
\docType{data}
\name{file_name}
\alias{file_name}
\title{Load GERDA Data}
\format{
An object of class \code{character} of length 1.
}
\usage{
file_name
}
\arguments{
\item{file_name}{A character string specifying the name of the file to load. For a list of available data, see \code{\link{gerda_data_list}}.}

\item{verbose}{A logical value indicating whether to print additional messages to the console. Default is FALSE.}
}
\value{
A tibble containing the loaded data.
}
\description{
This function loads GERDA data from a web source.
}
\examples{
\dontrun{
data_municipal_harm <- load_gerda_web("municipal_harm", verbose = TRUE)
}

}
\keyword{datasets}
