% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gerda_data_list.R
\name{gerda_data_list}
\alias{gerda_data_list}
\title{List of GERDA Data}
\usage{
gerda_data_list()
}
\value{
A tibble containing the available GERDA data with descriptions.
}
\description{
This function lists the available GERDA data sets.
}
\examples{
gerda_data_list()

}
