% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gerda_data_list.R
\name{gerda_data_list}
\alias{gerda_data_list}
\title{List of GERDA Data}
\usage{
gerda_data_list(print_table = TRUE)
}
\arguments{
\item{print_table}{A logical value indicating whether to print the table in the console (TRUE) or return the data as a tibble (FALSE). Default is TRUE.}
}
\value{
A tibble containing the available GERDA data with descriptions. When print_table = TRUE,
the function prints a formatted table to the console and invisibly returns the data tibble.
When print_table = FALSE, the function directly returns the data tibble.
}
\description{
This function lists the available GERDA data sets. The purpose of this function is to quickly provide a list of available data sets and their descriptions.
}
\examples{
gerda_data_list()

}
