% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_gerda_web.R
\name{load_gerda_web}
\alias{load_gerda_web}
\title{Load GERDA Data}
\usage{
load_gerda_web(file_name, verbose = FALSE)
}
\arguments{
\item{file_name}{A character string specifying the name of the file to load. For a list of available data, see \code{\link{gerda_data_list}}.}

\item{verbose}{A logical value indicating whether to print additional messages to the console. Default is FALSE.}
}
\value{
A tibble containing the loaded data.
}
\description{
This function loads GERDA data from a web source.
}
\examples{
\dontrun{
data_municipal_harm <- load_gerda_web("municipal_harm", verbose = TRUE)
}

}
