% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/party_crosswalk.R
\name{party_crosswalk}
\alias{party_crosswalk}
\title{GERDA}
\usage{
party_crosswalk(party_gerda, destination)
}
\arguments{
\item{party_gerda}{A character vector containing the GERDA party names to be converted.}

\item{destination}{The name of the column in the view_party table to map to.}
}
\value{
A vector with the mapped party names.
}
\description{
This function creates a crosswalk between parties and their corresponding names using the ParlGov view_party table. In cases where the party name is not found in the view_party table, the function returns NA. Note that this function should be run on GERDA party names, and will likely not work on other party naming schemes.
}
\examples{
party_crosswalk(c("cdu", "spd", "linke_pds", NA), "left_right")

}
